Script change
Hello guys,
The powershell that is available for download should be reviewed. After creating the application using the script (New-AadApp.ps1) and configuring the App Proxy settings, the option "Users and Groups" on Enterprise applications is removed. This is happening due the method that is used for service principal creation.  Instead of using New-AzureRmADServicePrincipal we should use New-AzureADServicePrincipal.